Combine select module and component files
