+++ app/src/main/java/de/tum/in/tumcampusapp/component/tumui/feedback/di/FeedbackComponent.kt
@@ -1,8 +1,12 @@
 package de.tum.`in`.tumcampusapp.component.tumui.feedback.di
 
+import dagger.Binds
 import dagger.BindsInstance
+import dagger.Module
 import dagger.Subcomponent
 import de.tum.`in`.tumcampusapp.component.tumui.feedback.FeedbackActivity
+import de.tum.`in`.tumcampusapp.component.tumui.feedback.FeedbackContract
+import de.tum.`in`.tumcampusapp.component.tumui.feedback.FeedbackPresenter
 
 @Subcomponent(modules = [FeedbackModule::class])
 interface FeedbackComponent {
@@ -18,5 +22,13 @@ interface FeedbackComponent {
         fun build(): FeedbackComponent
 
     }
+
+}
+
+@Module
+interface FeedbackModule {
+
+    @Binds
+    fun bindsFeedbackPresenter(impl: FeedbackPresenter): FeedbackContract.Presenter
 
 }
