--- app/src/main/java/de/tum/in/tumcampusapp/component/tumui/feedback/di/FeedbackModule.kt
@@ -1,14 +1,0 @@
-package de.tum.`in`.tumcampusapp.component.tumui.feedback.di
-
-import dagger.Binds
-import dagger.Module
-import de.tum.`in`.tumcampusapp.component.tumui.feedback.FeedbackContract
-import de.tum.`in`.tumcampusapp.component.tumui.feedback.FeedbackPresenter
-
-@Module
-interface FeedbackModule {
-
-    @Binds
-    fun bindsFeedbackPresenter(impl: FeedbackPresenter): FeedbackContract.Presenter
-
-}
