+++ app/src/main/java/de/tum/in/tumcampusapp/component/ui/ticket/di/TicketsComponent.kt
@@ -1,12 +1,16 @@
 package de.tum.`in`.tumcampusapp.component.ui.ticket.di
 
+import dagger.Binds
 import dagger.BindsInstance
+import dagger.Module
 import dagger.Subcomponent
+import de.tum.`in`.tumcampusapp.component.ui.ticket.EventsDownloadAction
 import de.tum.`in`.tumcampusapp.component.ui.ticket.activity.BuyTicketActivity
 import de.tum.`in`.tumcampusapp.component.ui.ticket.activity.EventsActivity
 import de.tum.`in`.tumcampusapp.component.ui.ticket.activity.ShowTicketActivity
 import de.tum.`in`.tumcampusapp.component.ui.ticket.activity.StripePaymentActivity
 import de.tum.`in`.tumcampusapp.component.ui.ticket.fragment.EventDetailsFragment
+import de.tum.`in`.tumcampusapp.service.DownloadWorker
 
 @Subcomponent(modules = [TicketsModule::class])
 interface TicketsComponent {
@@ -30,5 +34,13 @@ interface TicketsComponent {
         fun build(): TicketsComponent
 
     }
+
+}
+
+@Module
+interface TicketsModule {
+
+    @Binds
+    fun bindEventsDownloadAction(impl: EventsDownloadAction): DownloadWorker.Action
 
 }
