--- app/src/main/java/de/tum/in/tumcampusapp/component/ui/ticket/di/TicketsModule.kt
@@ -1,14 +1,0 @@
-package de.tum.`in`.tumcampusapp.component.ui.ticket.di
-
-import dagger.Binds
-import dagger.Module
-import de.tum.`in`.tumcampusapp.component.ui.ticket.EventsDownloadAction
-import de.tum.`in`.tumcampusapp.service.DownloadWorker
-
-@Module
-interface TicketsModule {
-
-    @Binds
-    fun bindEventsDownloadAction(impl: EventsDownloadAction): DownloadWorker.Action
-
-}
